use noeval by default (src, test, build)
